streamlit/requirements.txt
streamlit
pandas
snowflake-connector-python
snowflake-snowpark-python
matplotlib
google-generativeai
python-dotenv